updated for python3. implimented multithreading.
