+++ TODO.txt
@@ -1,6 +1,5 @@
 # TODO #
 
-* Multi-thread / Multi-Process requests
 * Increase variety for emails & passwords
 * option for random usernames
 * Auto pull the form data variables
